Refactoring some old code for better performance

+++ Database/GeoInfo/Currency/2021_10_26_095920_create_currency_table.php
@@ -1,5 +1,7 @@
 <?php
 
+use Caiocesar173\Utils\Enum\StatusEnum;
+
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -16,18 +18,23 @@ class CreateCurrencyTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create('currency', function(Blueprint $table) {
-            $table->uuid('id')->primary();
+		Schema::create('currency', function (Blueprint $table) {
+			$table->uuid('id')->primary();
 
 			$table->string('name', 255);
-			$table->string('code', 255)->nullable();
-			$table->string('currency_symbol', 255)->nullable();
-			$table->string('tld', 255)->nullable();
-			
-			$table->uuid('country');
-			$table->foreign('country')->references('id')->on('country')->onDelete('cascade');
+			$table->string('code', 5)->unique();
+			$table->string('symbol', 255);
+
+			if (env('UTILS_GEOLOC_ENABLE')) {
+				$table->uuid('country')->nullable();
+				$table->foreign('country')->references('id')->on('country')->onDelete('cascade');
+			} else
+				$table->string('country', 100)->nullable();
+
+			$table->enum('status', StatusEnum::lists())->default(StatusEnum::ACTIVE);
 			$table->timestamps();
-        });
+			$table->softDeletes();
+		});
 	}
 
 	/**
@@ -37,8 +44,8 @@ class CreateCurrencyTable extends Migration
 	 */
 	public function down()
 	{
-        Schema::disableForeignKeyConstraints();
-        Schema::drop('currency');
-        Schema::enableForeignKeyConstraints();
+		Schema::disableForeignKeyConstraints();
+		Schema::drop('currency');
+		Schema::enableForeignKeyConstraints();
 	}
 }
